Add and test main api
