--- index.js
@@ -1,5 +1,0 @@
-const { getQueries } = require("./src/index.js");
-
-module.exports = {
-  getQueries,
-};
